refactor(internal/config): move functions
Move functions from internal/config/config.go to other packages, so that
the config package only contains only type definitions for the
librarian.yaml configuration file.

Library.Fill is renamed to fillDefaults. It is also now called at the
start of generateLibrary.
