+++ internal/config/config.go
@@ -16,12 +16,6 @@
 // librarian.yaml configuration files.
 package config
 
-import (
-	"errors"
-)
-
-var errLibraryNotFound = errors.New("library not found")
-
 // Config represents a librarian.yaml configuration file.
 type Config struct {
 	// Language is the language for this workspace (go, python, rust).
@@ -141,71 +135,4 @@ type Channel struct {
 
 	// ServiceConfig is the path to the service config file.
 	ServiceConfig string `yaml:"service_config,omitempty"`
-}
-
-// LibraryByName returns a library with the given name.
-func (c *Config) LibraryByName(name string) (*Library, error) {
-	if c.Libraries == nil {
-		return nil, errLibraryNotFound
-	}
-
-	for _, library := range c.Libraries {
-		if library.Name == name {
-			return library, nil
-		}
-	}
-
-	return nil, errLibraryNotFound
-}
-
-// Fill populates empty library fields from the provided defaults.
-func (lib *Library) Fill(d *Default) {
-	if d == nil {
-		return
-	}
-	if lib.Output == "" {
-		lib.Output = d.Output
-	}
-	if lib.ReleaseLevel == "" {
-		lib.ReleaseLevel = d.ReleaseLevel
-	}
-	if lib.Transport == "" {
-		lib.Transport = d.Transport
-	}
-	if d.Rust != nil {
-		lib.fillRust(d)
-	}
-}
-
-// fillRust() populates empty fields in `lib.Rust` from the provided default.
-func (lib *Library) fillRust(d *Default) {
-	if lib.Rust == nil {
-		lib.Rust = &RustCrate{}
-	}
-	lib.Rust.PackageDependencies = mergePackageDependencies(
-		d.Rust.PackageDependencies,
-		lib.Rust.PackageDependencies,
-	)
-	if len(lib.Rust.DisabledRustdocWarnings) == 0 {
-		lib.Rust.DisabledRustdocWarnings = d.Rust.DisabledRustdocWarnings
-	}
-}
-
-// mergePackageDependencies merges default and library package dependencies,
-// with library dependencies taking precedence for duplicates.
-func mergePackageDependencies(defaults, lib []*RustPackageDependency) []*RustPackageDependency {
-	seen := make(map[string]bool)
-	var result []*RustPackageDependency
-	for _, dep := range lib {
-		seen[dep.Name] = true
-		result = append(result, dep)
-	}
-	for _, dep := range defaults {
-		if seen[dep.Name] {
-			continue
-		}
-		copied := *dep
-		result = append(result, &copied)
-	}
-	return result
 }
